# [Parquet] Allow opt-in prefetch metadata optimistically read in 1 request