--- a/extension/parquet/parquet_extension.cpp
+++ b/extension/parquet/parquet_extension.cpp
@@ -1754,6 +1754,9 @@ void ParquetExtension::Load(DuckDB &db) {
 	    "enable_geoparquet_conversion",
 	    "Attempt to decode/encode geometry data in/as GeoParquet files if the spatial extension is present.",
 	    LogicalType::BOOLEAN, Value::BOOLEAN(true));
+	config.AddExtensionOption("prefetch_metadata_bytes", "Optimistically prefetch that many bytes of parquet metadata.",
+	                          LogicalType::UBIGINT, Value(4088));
+	// Default is so that 4088 + 8 bytes that 4KB are always read on the first GET
 }
 
 std::string ParquetExtension::Name() {