--- a/extension/parquet/parquet_reader.cpp
+++ b/extension/parquet/parquet_reader.cpp
@@ -70,9 +70,9 @@ LoadMetadata(ClientContext &context, Allocator &allocator, FileHandle &file_hand
 		throw InvalidInputException("File '%s' too small to be a Parquet file", file_handle.path);
 	}
 
-	Value value;
+	Value value = 0;
 	context.TryGetCurrentSetting("prefetch_metadata_bytes", value);
-	const idx_t prefetch_metadata_bytes_option = UBigIntValue::Get(value);
+	const idx_t prefetch_metadata_bytes_option = value.GetValue<uint64_t>();
 
 	const idx_t requested_footer_size = std::min((idx_t)file_size, prefetch_metadata_bytes_option + 8);