--- a/test/sql/copy/parquet/test_parquet_force_download.test
+++ b/test/sql/copy/parquet/test_parquet_force_download.test
@@ -9,6 +9,9 @@ require httpfs
 require tpch
 
 statement ok
+SET prefetch_metadata_bytes = 0;
+
+statement ok
 SET force_download=true;
 
 query II