@carlopi 2024-12-04
Need to add [de]serialization code, see https://github.com/duckdb/duckdb/pull/15135