@carlopi 2025-05-15
Outdated after https://github.com/duckdb/duckdb/pull/17300